+++ src/function/lock.js
@@ -3,16 +3,17 @@ import { assertFunction } from '../util/assertions';
 /**
  * Locks the number of arguments accepted by the function.
  *
- * @extra If not passed, `n` will be the length of the function. Note that when
- * using with other methods that bind arguments such as `Function#bind` or
- * `partial` this method needs to be called first in the chain to prevent bound
- * arguments from being passed.
+ * @extra Note that when using with other methods that bind arguments such as
+ * `Function#bind` or `partial` this method needs to be called first in the
+ * chain to prevent bound arguments from being passed.
+ *
+ * @param {Function} fn - The function to lock.
+ * @param {number} [n] - The arity (number of arguments) to lock the function
+ *   to. If not passed it will be the `.length` of the function.
  *
  * @example
  *
  *   logArgs.lock(2)(1,2,3) -> logs 1,2
- *
- * @param {number} [n]
  *
  **/
 export default function lock(fn, n = fn.length) {